refactor: типы ингредиентов вынесены
