+++ src/components/burger-constructor/bun/bun-bottom.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import { ConstructorElement } from '@ya.praktikum/react-developer-burger-ui-components';
 
-import ingredientPropTypes from '../../prop-types/ingredient-prop-types';
+import ingredientPropTypes from '../../../types/ingredient-prop-types';
 import styles from './bun.module.css';
 
 function BunBottom({ bun, text = 'Добавьте булку' }) {
