+++ src/services/selectors/order.js
@@ -1,4 +1,5 @@
 import { createSelector } from 'reselect';
+import { BUN } from '../../types/ingredient-types';
 
 export const getStuffing = createSelector(
   (store) => store.order.ingredients,
@@ -21,7 +22,7 @@ export const getCountStuffing = createSelector(
   (store) => store.order.ingredients,
   (_, ingredient) => ingredient,
   (ingredients, { _id, type }) => {
-    if (type === 'bun') {
+    if (type === BUN) {
       return ingredients.bun?._id === _id ? 2 : 0;
     }
     return ingredients.stuffing.filter((item) => item._id === _id).length;
