+++ src/types/ingredient-types.js
@@ -1,0 +1,3 @@
+export const BUN = 'bun';
+export const MAIN = 'main';
+export const SAUCE = 'sauce';
